Add: explicit declare model relationship in backend; Change: build instructions, badge and progress in README
